fix(0.9.5): mixin regression — DEFAULT_SETTINGS et al. unresolved
v0.9.0/v0.9.1 splits moved methods into separate modules but their
bodies kept referencing module-level names by bare identifier. /ask
crashed with 'name DEFAULT_SETTINGS is not defined' (and 16 sibling
errors).

Two new shared modules:
* _agent/_types.py     — SearchPlan, SearchPolicy, helpers
* _catalog/_constants.py — DEFAULT_SETTINGS, SOURCE_PRIORITY, helpers

Both agent.py + catalog.py re-export (public API preserved); each
mixin imports what it references. Verified: 0 unresolved references
across all 13 mixin files.
