+++ README.md
@@ -25,6 +25,7 @@ Results from all agents are **merged** by an orchestrator using LLM reasoning, t
 - Write approved metadata back to the database as `COMMENT ON TABLE/VIEW/COLUMN` (write-back support)
 
 Recent release notes:
+- `v0.9.5`: **Fix mixin regression (`name 'DEFAULT_SETTINGS' is not defined` and 16 sibling errors).** v0.9.0/v0.9.1 mixin splits left module-level names (`SearchPlan`, `_input_token_budget_for`, `DEFAULT_SETTINGS`, `SOURCE_PRIORITY` …) referenced by bare name from mixin method bodies — but each mixin was now its own module. Fix: two new shared modules `_agent/_types.py` and `_catalog/_constants.py` own the dataclasses + helpers + constants; `agent.py`, `catalog.py`, and every mixin import from there. Public API preserved (`from amx.search.agent import SearchPlan` still works via re-export). Reproducer (`/ask` after a clean install) now passes.
 - `v0.9.4`: **`execute_analyze_run` extracted into 3 phase helpers (S4 refactor).** The `/run` entry point was a 600-LOC procedural script juggling 12 local variables across 3 exception handlers; three different bugs surfaced there in recent conversations. v0.9.4 lifts the three largest chunks into a new `amx/cli_support/commands/_analyze/` package: `run_loop.py` (per-schema orchestration loop), `run_summary.py` (post-loop summary + apply branch), `interrupt.py` (KeyboardInterrupt body decision). execute_analyze_run is now 390 LOC (-35%); analyze_flow.py: 1099 → 877 LOC. Each phase is independently testable; exception handlers no longer race the main path for shared variables.
 - `v0.9.3`: **Slash commands collapsed into a single registry (S5 refactor).** Before, each command had to be listed in 4 places in `session.py` (autocomplete catalog, `*_cmd_heads` frozensets, namespace help, dispatch chain). v0.9.3 introduces `amx/cli_support/slash_commands.py` — a 301-LOC declarative registry of `SlashCommand` dataclasses; the autocomplete catalog (5-line adapter) + dispatch heads (single helper call) now derive from it. session.py: 1283 → 1125 LOC. Drift between sources is now structurally impossible. Side-effect: the `/tls` command, previously in autocomplete but missing from dispatch, is now correctly routed.
 - `v0.9.2`: **`Orchestrator.process_table` god-method extracted into `TableProcessor` (S3 refactor).** The 281-LOC method grew four filter chains plus an agent loop plus three apply branches; each new feature (Phase 2 dedup, Column scope) layered another filter on top. v0.9.2 extracts the flow into `amx/agents/_orchestrator/table_processor.py` (12 phase methods, none over 77 LOC). `Orchestrator.process_table` is now 25 lines — a thin delegator. Public signature unchanged. Each filter is now unit-testable in isolation; the three apply branches no longer compete for context. orchestrator.py: 1719 → 1463 LOC.
