+++ amx/search/_agent/answering.py
@@ -22,6 +22,14 @@ import json
 from dataclasses import asdict
 from typing import Any
 
+from amx.search._agent._types import (
+    SearchActionSuggestion,
+    SearchPlan,
+    SearchPolicy,
+    _input_token_budget_for,
+    _question_language_hint,
+    _trim_rows_to_token_budget,
+)
 from amx.utils.logging import get_logger
 
 log = get_logger("search.agent.answering")
