+++ amx/search/_agent/resolution.py
@@ -27,9 +27,8 @@ from __future__ import annotations
 import re
 from typing import Any
 
+from amx.search._agent._types import ResolvedTarget, SearchPlan
 from amx.utils.logging import get_logger
-
-ResolvedTarget = Any  # Forward reference to dataclass in agent.py.
 
 log = get_logger("search.agent.resolution")
