+++ amx/search/_catalog/settings.py
@@ -14,6 +14,7 @@ from __future__ import annotations
 
 from typing import Any
 
+from amx.search._catalog._constants import DEFAULT_SETTINGS
 from amx.utils.logging import get_logger
 
 log = get_logger("search.catalog.settings")
